+++ .github/CHANGELOG.md
@@ -1,12 +1,40 @@
-### 0.6.15
+### 0.7.16
+
+**Features**
+
+1. Check for updates..., and auto update when update available.(Still need signature...:cry:)
+
+2. Insert Image: ( In edit menu )
+
+   - absolute path
+
+   - relative path
+
+   - Upload Image to cloud
+
+3. Add file icons to languages when create code block or change language in code block.
 
 **Bug fix**
 
 1. It's hard to focus the input in code fence.
+
 2. When input the language in code block, click the language item will not cause hide the float box.
+
 3. other bugs in code block.
 
-### 0.6.12
+4. Windows user can not use open with feature.
+
+5. The menu disapear in Linux sysyem.
+
+6. Fix the bug that the language highlight disapear when open markdown file with code block.
+
+**Optimization**
+
+1. allow user to change install directory on windows.
+
+2. Show notification when output HTML and PDF successfully.
+
+### 0.6.14
 
 **Features**
 
@@ -25,6 +53,7 @@
 **Bug fix**
 
 - Fix bug when search key in code block will cause the search input lose focus.
+
 - Fix the bug the editor will lose cursor after input Chinese.
 
 ### 0.5.2